dns-dir-review-nits-part
Covers this part of the review

And I also have some editorial comments:

In section 4, “The contact details of the IT/InfoSec team to report mis-classified DNS filtering. This field is structured as an array of contact URIs (e.g., tel, sips, https). At least one contact URI MUST be included. This field is mandatory.” It is necessary to reference RFCs to “tel, sips, https”.

In section 5.3, there is an in-paragraph long space breaking “If a DNS client has enabled opportunistic privacy profile (Section 5 of [RFC8310]) for DoT, the DNS client will either fall back to an...” ...and “encrypted connection without authenticating the DNS server...”.

In section 5.3, the first action is described as “Verify the field contains valid JSON.” which is the only segment using a verb to describe the very action. I think it would be better to align all the action description wording.